+++ src/main/java/nl/_42/heph/lazy/LazyEntityReference.java
@@ -1,4 +1,4 @@
-package nl._42.heph;
+package nl._42.heph.lazy;
 
 import java.util.function.Consumer;
 import java.util.function.Supplier;
@@ -11,7 +11,7 @@ import org.springframework.data.domain.Persistable;
  * reference will be resolved and set using the the setter consumer.
  * @param <T> classtype of the entity that is being resolved
  */
-public class LazyEntityReference<T extends Persistable> extends AbstractLazyEntity<T,T>{
+public class LazyEntityReference<T extends Persistable> extends AbstractLazyEntity<T,T> {
 
     public LazyEntityReference(Supplier<T> getter, Consumer<T> setter, Supplier<T> reference) {
         super(getter, setter, reference);