+++ src/test/java/nl/_42/heph/builder/PersonFixturesTest.java
@@ -3,40 +3,36 @@ package nl._42.heph.builder;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
 
-import java.io.Serializable;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Supplier;
 
-import javax.persistence.EntityNotFoundException;
-
-import nl._42.heph.AbstractBuildCommand;
 import nl._42.heph.domain.Organization;
 import nl._42.heph.domain.OrganizationRepository;
 import nl._42.heph.domain.Person;
-import nl._42.heph.domain.PersonRepository;
 import nl._42.heph.shared.AbstractSpringTest;
 
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.JpaRepository;
 
-public class PersonBuilderTest extends AbstractSpringTest {
+public class PersonFixturesTest extends AbstractSpringTest {
 
     @Autowired
-    private PersonRepository personRepository;
+    private PersonFixturesRepository personRepository;
     @Autowired
     private OrganizationRepository organizationRepository;
     @Autowired
-    private PersonBuilder personBuilder;
+    private PersonFixtures personFixtures;
 
     @Test
     public void checkLazyCreation() {
         // Create a person with the same name by hand
         Person person = new Person();
-        person.setName(PersonBuilder.EXPECTED_NAME);
+        person.setName(PersonFixtures.EXPECTED_NAME);
         personRepository.save(person);
 
-        personBuilder.base().create();
+        personFixtures.base().create();
 
         List<Organization> organizations = organizationRepository.findAll();
         assertEquals(0, organizations.size());
@@ -45,8 +41,8 @@ public class PersonBuilderTest extends AbstractSpringTest {
     @Test
     public void copy() {
         final String expectedName = "Somebody else";
-        Person aPerson = personBuilder.sjaak();
-        Person theOtherPerson = personBuilder.copy(aPerson)
+        Person aPerson = personFixtures.sjaak();
+        Person theOtherPerson = personFixtures.copy(aPerson)
                 .withName(expectedName)
                 .create();
         assertNotSame(aPerson.getId(), theOtherPerson.getId());
@@ -54,13 +50,26 @@ public class PersonBuilderTest extends AbstractSpringTest {
     }
 
     @Test
+    public void valueStore() {
+        final AtomicInteger callbacksMade = new AtomicInteger(0);
+
+        Person aPerson = personFixtures.base()
+                .withName("aName")
+                .withCallbackFunction((person) -> person.setName("aName #" + callbacksMade.incrementAndGet()))
+                .create();
+
+        assertEquals(1, callbacksMade.get());
+        assertEquals("aName #1", aPerson.getName());
+    }
+
+    @Test
     public void createWithoutSpringContext() {
-        assertCreationWithoutSpringContext(new PersonBuilder().base()::create);
+        assertCreationWithoutSpringContext(new PersonFixtures().base()::create);
     }
 
     @Test
     public void constructWithoutSpringContext() {
-        assertCreationWithoutSpringContext(new PersonBuilder().base()::construct);
+        assertCreationWithoutSpringContext(new PersonFixtures().base()::construct);
     }
 
     private void assertCreationWithoutSpringContext(Supplier<Person> personSupplier) {
@@ -68,8 +77,8 @@ public class PersonBuilderTest extends AbstractSpringTest {
         // in PersonBuilder to be manually constructed as well,
         // besides being @Autowired.
         Person person = personSupplier.get();
-        assertEquals(PersonBuilder.EXPECTED_NAME, person.getName());
-        assertEquals(OrganizationBuilder.EXPECTED_NAME, person.getOrganization().getName());
+        assertEquals(PersonFixtures.EXPECTED_NAME, person.getName());
+        assertEquals(OrganizationFixtures.EXPECTED_NAME, person.getOrganization().getName());
     }
 
 }