+++ src/test/java/nl/_42/heph/builder/UnsavableEntity.java
@@ -1,0 +1,22 @@
+package nl._42.heph.builder;
+
+import javax.persistence.Entity;
+
+import nl._42.heph.shared.AbstractEntity;
+
+/**
+ * Entity for which no repository exists.
+ */
+@Entity
+public class UnsavableEntity extends AbstractEntity {
+
+    private String version;
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
+    }
+}