+++ src/test/java/nl/_42/heph/builder/WorkspaceFixturesTest.java
@@ -1,0 +1,4 @@
+package nl._42.heph.builder;
+
+public class WorkspaceFixturesTest {
+}